Refactoring and docs.
